+++ tests/integration/causedby-annotation/test.php
@@ -19,3 +19,23 @@ function testCausedBy() {
 function passParamThrough( $par ) {
 }
 echo passParamThrough( $_GET['a'] ); // There must be a caused-by line pointing to the annotation
+
+
+/**
+ * @param-taint $par exec_html
+ */
+function execArg( $par ) {
+	echo $par;
+}
+execArg( $_GET['a'] ); // Must have the annotation in caused-by, and NOT line 28
+
+/**
+ * @param-taint $par html
+ */
+function preserveHtmlAnnotatedAddYes( $par ) {
+	$ret = $par;
+	$ret .= $_GET['a'];
+	return $ret;
+}
+echo preserveHtmlAnnotatedAddYes( $_GET['a'] ); // Must have lines 38, 37 and annotation (not 36)
+shell_exec( preserveHtmlAnnotatedAddYes( $_GET['a'] ) ); // Must have lines 38, 37 (not 36 and annotation)